feat: modify imports to organise queries
