+++ package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/ivanleekk/cashshare-telegram#readme",
   "dependencies": {
+    "@dotenvx/dotenvx": "^1.14.2",
     "@prisma/client": "^5.20.0",
     "@vitest/coverage-v8": "^2.1.1",
     "aws-lambda": "^1.0.7",
