+++ src/handlers/balanceHandler/balanceHandler.ts
@@ -1,5 +1,6 @@
-import { findUser_byUsername, sendMessage } from "../../utils/utils";
+import { sendMessage } from "../../utils/telegramUtils";
 import { PrismaClient } from "@prisma/client";
+import {findUser_byUsername} from "../../utils/prisma/prismaUserUtils";
 
 const prisma = new PrismaClient({});
