+++ src/handlers/globalHandler.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { APIGatewayProxyEvent, Context } from 'aws-lambda';
 import { globalHandler } from './globalHandler';
-import { sendMessage } from '../utils/utils';
+import { sendMessage } from '../utils/telegramUtils';
 import { startHandler } from './startHandler/startHandler';
 import { addHandler } from './addHandler/addHandler';
 import { groupBalanceHandler, individualBalanceHandler } from './balanceHandler/balanceHandler';
