+++ src/utils/prisma/prismaUserUtils.ts
@@ -1,0 +1,9 @@
+import {prisma} from "../../../libs/prisma";
+
+export async function findUser_byUsername(username: string) {
+    return prisma.user.findUnique({
+        where: {
+            username: username
+        }
+    });
+}
